Semi Big Update
Game is feature complete with word checking integration.         Basic SQL integration, now if you login and exit out of the app, it will automatically log you back in.'
If the app loses focus, meaning you press home and go do something else on the phone, then it will disconnect in the background and reconnect to server
Point system for client and server is in place but it is not exposed to the user as of right now.